requirements.txt
allure-pytest-bdd==2.13.1
allure-python-commons==2.13.1
async-generator==1.10
attrs==21.2.0
bdd==0.2
behave==1.2.6
behave2cucumber==1.0.3
behave2cucumber-py3==1.0.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
configparser==5.1.0
cryptography==35.0.0
driver==0.1.0
exceptiongroup==1.1.1
glob2==0.7
h11==0.12.0
idna==3.3
iniconfig==1.1.1
lml==0.1.0
Mako==1.1.5
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.2
parse==1.19.0
parse-type==0.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
pytest==6.2.5
pytest-bdd==6.1.1
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-parallel==0.1.1
python-dotenv==1.0.0
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tblib==1.7.0
toml==0.10.2
tqdm==4.65.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.5.0
urllib3==1.26.7
urllib3-secure-extra==0.1.0
webdriver-manager==3.8.5
wsproto==1.0.0
